Add logic to handle links and refactor project data import
The update introduces functionality to handle 'links' into 'project-json.lua'. In the 'binding.gyp.mjs' file, the import of project data was refactored by moving the 'exec' function to a new 'project.mjs' file. This improves the data extraction and makes the code cleaner. A 'toArray' utility function was also added to 'utils/index.mjs'. Finally, the 'postinstall' script was added to 'package.json' to run 'tsc'.